Remove unhelpful axi wrapper runt tests and cocotb tests (#2515)
This PR gets rid of the yxi-based wrapper runt tests mentioned in #2233.

These files tried to snapshot "steps" in the process of going from a
calyx program to AXI based cocotb testing, but the process for updating
was manual (updating snapshots one at a time then copying the new
`*.expect` to the correct `*.file-suffix`) and error prone. (Forgetting
to update a `*.file-suffix` would just leave the tests passing). Once
work starts up on AXI again I do think something in the spirit of these
tests would be useful to have. The cocotb axi tests working is a good
starting point to making sure that the AXI wrappers are [at least
partially] correct.

With the changes introduced in #2516 being not well tested, I think it
makes sense to get rid of these tests, as long as we want to merge those
changes as well.

Some more context in
[zulip](https://calyx.zulipchat.com/#narrow/channel/445339-calyx-yxi/topic/AXI.20Controller/near/528563512)