[Fud2] Extract execution timing from Ninja (#2525)
A quick extension to `fud2` which will output a csv of execution timing
when the `--csv` argument is given with a filename. This is similar to
functionality in the original `fud`.

This works by extracting the timing info placed in the `.ninja_log`
which reports the timestamp in milliseconds (with zero being the start
of the overall build) at the beginning and end of the build command for
the given file. I'm not certain how precise this information is, but
this should at least give a reasonable ballpark estimate for how long
the individual parts of an execution are taking. Unfortunately, while
the `.ninja_log` file seems stable in its construction, it is not
formally documented anywhere and could potentially change down the line
which would break this functionality.