[Cider] Thread pausing and evaluation polices (#2531)
This PR introduces functionality to the program counter which can allow
for individual control nodes to be paused during execution and provides
three different polices for how this should be done.

The policy decides whether or not new nodes added to the program should
be paused and what nodes to unpause when then execution is stalled. The
current polices are as follows:
- `default`: All new nodes are always enabled. When execution stalls it
reports an error.
- `random`: New nodes are randomly enabled with a 50/50 chance. When
execution stalls, all paused nodes have a 50/50 shot of being enabled.
- `random_sequential`: When new nodes are added, exactly one is enabled.
When execution stalls, exactly one of the paused nodes are enabled.

The current stall detection is very rudimentary. It merely checks that
something somewhere in the execution has made updates or alterations to
the program state, so it is fairly easy to write groups that will never
terminate which are not detected as making no progress. In the future, a
more advanced analysis could potentially check for loops in the program
state, but that is left for future efforts.