[repo] Bump rust version and dependencies (#2532)
I've bumped the rust version to `1.88` and also updated the versions of
our rust dependencies. As this updates `clippy`, that means a bunch of
new lints are active now, so I had to go and correct those as well. The
vast majority are format string related.

Unfortunately that makes for a bit of a big PR.