Switch `morty` dependency to git (#2535)
The released version as of this writing (0.9.0) indirectly depends on an
old version of nom that is now emitting a worrisome
future-incompatibility warning. However, the dependencies are fresh
enough on git that just using that makes the warning go away.

It is slightly worrisome to be depending on a specific git commit, but
maybe less bad than the alternative?