Actually install calyx-py
The queues generator, for example, is included in the test suite and
just expects this to be installed globally. So let's install it
globally! It would be even nicer to, someday in the future, require all
these Python things to work in their own virtualenvs, or otherwise run
with uv without requiring manual dependency setup.