Try an explicit install of calyx-py
It turns out that workspace dependencies get used "via path" in uv,
instead of being copied into the virtualenv's lib directory:
https://github.com/astral-sh/uv/issues/9362

Doing this explicit install of calyx-py *after* fud makes sure it's
installed in the conventional (non-editable) way. This matches how
things worked when we were using flit to install.

Also reverts f4167f402e45b41987793e2e6c88f9017d732299, which was a
different workaround for the same issue. But it wasn't completely
perfect because it didn't match the old behavior for frontends that
imported the `gen_*` scripts as libraries (which relay does, for
example).