Add end to end Calyx-AXI wrapper cocotb simulation runt test (#2540)
This gives us a single runt test that goes from a calyx vector add to
its expected output after being simulated with cocotb.

This also changes some wire names in the compute-read-write/static AXI
wrapper generator that aligns with what cocotb expects. Now the dynamic
and static versions are in sync.
Can replace a lot of what was removed in #2515.

Pleasantly surprised to find that the static generator still simulates
correctly