Dockerfile: Switch from flit to uv (#2539)
Following up on #2534, we apparently now _need_ to use uv to do this
install to take advantage of the workspace (i.e., fud's dependency on
calyx-py). So I have switched the Dockerfile to use uv instead of flit
everywhere!

I am hoping this resolves the CI failure showing up in #2530.