[Cider] Some par bugfixes and assorted changes to some of the generators (#2545)
Apologies for the mishmash of stuff in one PR but it's probably easier
to manage than splitting it up into a bunch of tiny ones.

This contains several bugfixes related to par blocks and threads
- One major issue was par blocks not properly restoring the parent
thread on exit in cases when the parent's thread was determined
implicitly. This lead to subcomponents getting their thread ID stuck to
the first thread that happened to invoke them, consequently leading to
incorrect data race detection on some programs which invoke the same
component from different threads.
- Another issue with par handling is that empty par blocks would cause
entire threads of execution to disappear because the processing logic
assumed they would always have some children. These are now handled like
empty seqs and are simply moved past immediately, akin to an empty
group.

Additionally
- Finally added a flag to disable thread memoization, which you
shouldn't do because it has a very bad impact on performance.
- Added an additional test case generator for diagonal matrix
computation
- Also made some changes to the python generators used to make larger
cider test cases.