[Fud2] Tweak error message to list states (#2546)
Closes #2206 

A small change to the could not find path error message. It now lists
the set of sources and the set of destinations that it failed to find a
path for. Probably should do something better for the multi-input
version but I don't know enough about that to properly speculate. Only
other thing of note is that I changed the plan function to take a
reference to the request rather than owning it which results in one
clone of a pathbuf which I think is acceptable.

Now when running a nonsense command, such as going from dat to
cider-debug we get the following error message:
```
Error: could not find path from {dat} to {cider-debug}
```