[Profiler] Ordering control groups using descriptor information (#2527)
This PR addresses the problem with control groups not appearing
correctly in order on the call stack (the first part of next steps
listed in #2522 ).

## Implementation strategy

The high level strategy is to use path descriptors to identify the
control node ancestry of groups and control nodes.

### During compilation

The `uniquefy-enables` pass now also outputs a mapping between control
node descriptors to the list of `@pos` Ids of the control node, when a
file is provided in the `path-descriptor-json` argument. It also makes
unique versions of comb groups, and keeps track of their par track IDs
and their descriptors as well.

### During VCD processing

We produce a mapping from group enables (unique groups for each enable)
to their descriptors, and another mapping from control groups to their
descriptors (using `@pos` information emitted from the
`uniquefy-enables` and `tdcc` passes).

The algorithm to create a control-group-included trace starts with the
original trace (no control groups). For each stack active in a cycle `i`
(there could be multiple bc parallelism), we create a new control stack
that contains the same elements as the stack, but with control groups
added between a (user-defined) component cell and the group enabled from
the cell's control block. The control groups that are active and are an
"ancestor" of the group enabled (i.e. the control groups' path
descriptors are prefixes of the group's path descriptor) are added,
based on the alphabetical order of path descriptors. (If there are no
groups active from the cell's control, all active control groups are
added)

## Interesting edge case

An interesting edge case/bug I ran into today was when control groups
that were active were in a control register transition cycle while _in a
group was executing in a different par arm_. To explain with a simple
example, suppose we had the following control block in `main`:

```
par {
  a;
  if cond.out { b; }
}
```

Assuming that no groups nor control nodes get converted to static and
`cond.out == 1'b1`, we will end up with two control groups, one to
manage the `par` (let's call this `par_group`) and the other to manage
the `if` (let's call this `if_group`, it also uses an FSM).

In cycle 1, `a` will be enabled, while `if_group` will take one cycle to
update its FSM to the state where `b` is enabled (so `b` is not active
yet). So, the non-control stacks for cycle 1 looks something like: `[
main, a ]`. While `par_group` and `if_group` are both active in cycle 1,
if we straight up use the algorithm listed above, there will only be one
control stack that looks like: `[main, par_group, a]`. This makes it
seem like `if_group` is not active at all.

To solve this edge case, I added an extra helper function(s) (currently
named `TraceData._create_stacks_for_missed_control_groups()`) that
create extra stacks for any control groups that are "missing" in this
way.

## Caution/Should be fixed in the future...

This PR steps even deeper in the conflation of source mappings and
unique identifiers for control nodes/groups by using the `@pos`
attribute. Specifically, the `uniquefy-enables` pass maps the path
descriptor of a control node (which should uniquely identify the control
node) to all `@pos` attribute IDs of the same control node. The
attributes are passed down through the compiler to connect the control
group that structurally implements the control node is ultimately
associated with the path descriptor of the original control node.

This problem was also listed in #2522 but got worse in this PR (it was
necessary for a first pass), and should be addressed in the near future.

---------

Co-authored-by: Griffin Berlstein <griffin@berlste.in>