[Compiler] Add a pass to propagate protected attributes (#2548)
Closes #2538.

Adds a pass which propagates `@protected` attributes from a primitive
definition to their instances. I placed it at the beginning of the
`pre-opt` group of passes but could see it being placed elsewhere.

I originally thought about doing this as part of the conversion from AST
to IR but figured it would be better to make it a pass, though that is
definitely more code on the whole.