[Profiler] Refactoring: Implementing the trace with a dense list (#2554)
Addressing the second bullet point in @sampsyo 's [helpful list of
refactoring
suggestions](https://github.com/calyxir/calyx/pull/2461#pullrequestreview-2865047008)!

The trace was previously implemented as a dictionary mapping cycles to
the set of "stacks" active that cycle. In this PR, we replace these
dictionaries with a `PTrace` class which functionally works like a
dictionary, but is implemented by a dense list.