[Cider] Performance tuning (#2557)
Collecting the work I've been doing on and off for the past few weeks.
Made a bunch of small changes across different parts of the cider code
base to improve execution time when running with data race detection
turned on. There's no silver bullet trick here sadly, so I don't have
anything super interesting to discuss here.

For some of the larger examples this seems to approximately cut the
runtime in half and in general performance in the non-checked case has
seen minor improvements as well.