[Cider] Add a flag to allow entangled memories to be passed as a file (#2560)
A small ergonomics tweak to make working with test cases which require
larger numbers of entangled memories. The `--entangle-file` flag can now
indicate a file, with one memory name on each line, which should be
entangled in the same way as a list provided via the CLI.

Note that because of the way fud2 works this currently presents a
usability issue as the file names must either be given in absolute paths
or prepend a `../` in order to be found during runtime. I don't have a
tidy fix for this yet, short of embedding knowledge of entangled files
into the fud2 stage. At any rate, this is something to fix later.