[Cider] Proper source printing (#2563)
Small set of changes to allow cider to print the source lines indicated
by the sourcemap when they are available during execution. If unable to
find the file it will print an error message and then continue
execution. Currently this error message is not suppressed by the quiet
setting. The correct thing to do would probably be to display it through
logging as a warning to allow it to be suppressed, but I leave that for
another time