[Profiler] First pass for Dahlia-level profiling (#2567)
We support profiling at the Dahlia level!

In order to do so, we added:
- Updates to the `fileinfo_emitter` tool to handle Dahlia code. For "ADL
variable name", the entire corresponding line of Dahlia is recorded.
- Created flamegraphs and timeline views at the Dahlia statement level
- Generalized the flamegraph and timeline view emitting tools so it
accommodates both Dahlia and Calyx program structures.

## Next steps

I'd like to improve the flame graphs and timeline views to contain some
nesting. For example, statements that happen within a for loop should
show up as "children" of the for loop. This should hopefully make events
easier to see and understand, but we'll see!