[fud2] Json Parser for Serialized Plans  (#2564)
This serializes fud2 using JSON and a language called flang. See #2555
for more details. In short this PR adds one new planner: `json`. It
reads a plan from stdin specified in the JSON form of flang.

Additionally the PR adds one new mode: `json-plan`, taking the plan from
the current planner and writing it to stdout in JSON syntax of flang.