[Compiler] Expand the source info table with initial information about state (#2575)
This PR updates the source info table with three new optional tables:
- the `memory_locations` table assigns a numeric identifier to named
memory locations, either registers or memory addresses
- the `variable_assignments` table assigns a numeric identifier to a
collection of pair of source variable names and memory location ids
- the `position_state_table` maps from position ids to the variable
assignment sets that are active during the portion of the program
covered by the position tag.

These tables may be omitted from the source info block if only the file
and position information is needed. However, as the tables are
interdependent, they must all be present when included.

Next steps will be allowing Cider to use the map to parse print requests
for variables.