[Profiler] Dahlia blocks, first pass for primitives on timeline view (#2579)
This PR improves the profiler's timeline view. The timeline view is now
given two additional features:
1. Adds blocks (`for`, `while`, and `if`) and nesting to Dahlia program
views.
We nest statements within the `while`, `for`, and `if` blocks that they
belong in, and blocks last the total duration of their statements.
2. Adds active durations of Calyx primitives
In the Dahlia timeline view, we now have a dropdown `Calyx Primitives`
that contains all the Calyx primitives used in the program (organized by
primitive type). In the Calyx timeline view, we have a `Primitives`
dropdown for each Calyx component cell that shows the primitives used in
that Calyx component.

Here's a screenshot demonstrating the new entries in the Dahlia timeline
view (for `linear-algebra-bicg.fuse`):

<img width="490" height="648" alt="image"
src="https://github.com/user-attachments/assets/1df13248-e933-4c7b-b723-a0f082ec0809"
/>

# Next steps

We may change where to put the Calyx primitives - it might be insightful
to see what specific statements trigger a Calyx primitive instead of
having a dropdown for all Calyx primitives.