Refresh Python dependencies (#2594)
The main goal here was to do a `uv lock --upgrade` to get newer versions
of all our Python dependencies. (The original purpose was to hopefully
avoid a dependency error that @ayakayorihiro was getting on her
machine.) This required some fiddling with constraints to make Pynq
work, but nothing major.

Fingers crossed that this disrupts nothing in CI.